pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dailyplan"
version = "0.2.0"
description = "Generate daily task plans as iCalendar (.ics) files"
readme = "README.md"
authors = [{ name = "Monetly" }]
license = { text = "MIT" }
dependencies = ["pyyaml"]

[project.scripts]
myplan = "dailyplan.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
